the version of imagemagick currently installed chokes on XMP profiles in a lot of our images. it prints out a warning that ultimately results in a 500.
by suppressing the warning, the processing continues.  longer term fix is probably to upgrade to imagemagick 7.1.x
which on the mac at least doesn't seem to be printing errors.  this will require burning a new node/alpine base image
though, so for now it's easier to patch this at the app level